requirements.txt
asgiref==3.8.1
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
mysqlclient==2.2.7
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
